common_config

Module config

Constants

ENV_LIST_SEP

Separator for list values in environment variables. For example, localhost:3001,localhost:3002,localhost:3003.

ENV_VAR_SEP

Separator for environment variables. For example, DATANODE__STORAGE__MANIFEST__CHECKPOINT_MARGIN.

Traits

Configurable

Configuration trait defines the common interface for configuration that can be loaded from multiple sources and serialized to TOML.